Fix require statements as a result of the http1 -> http/1 move.
Introduces '../../' require statements.
